requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Django==5.2.5
django-cors-headers==3.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
pip==25.1.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
schedule==1.2.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
Werkzeug==3.1.3
